requirements.txt
# core daily-dragon dependencies
fastapi
pydantic
python-dotenv
boto3
botocore

# for AWS Lambda deployment
mangum

pytest
pytest-cov
httpx

# for running FastAPI app locally
uvicorn